Broadcast device and user identity updates
This patch adds a way for users to listen for devices and user
identities that were added or updated in our store.

This becomes useful if and when users try to implement a page which
displays a list of devices a user has, or if they want to warn about
newly added devices.